+++ src/collective/taxonomy/__init__.py
@@ -1,2 +1,6 @@
+# -*- coding: utf-8 -*-
+
 PATH_SEPARATOR = u"\u241F"
 LEGACY_PATH_SEPARATOR = u"/"
+PRETTY_PATH_SEPARATOR = u' » '
+NODE = u"#NODE"